Fixed Collider Value Pointers for Collider + Collision Manager Setup
- Colliders now have pointers that point to memory address of things like the objects x, y, width, height and radius (CORRECTLY NOW).
- Added generic functions for BoxVBox Collision, CircleVCircle Collision and BoxVCircle Collision (Logic not in them yet)

+++ EAT Assessment Project/BoxCollider.cpp
@@ -1,6 +1,6 @@
 #include "BoxCollider.h"
 
-BoxCollider::BoxCollider(int &x, int &y, int &width, int &height)
+BoxCollider::BoxCollider(float &x, float &y, int &width, int &height)
 {
 	this->x = &x;
 	this->y = &y;
